remove spectate.core
