+++ .python-version
@@ -1,0 +1,1 @@
+spectate
