+++ spectate/__init__.py
@@ -1,3 +1,1 @@
-__version__ = "0.4.1"  # evaluated in setup.py
-
-from .core import *  # noqa
+__version__ = "0.5.0"  # evaluated in setup.py
