+++ spectate/mvc.py
@@ -1,0 +1,3 @@
+from .base import *  # noqa
+from .events import *  # noqa
+from .models import *  # noqa
