--- spectate/mvc/__init__.py
@@ -1,10 +1,0 @@
-import sys
-
-if sys.version_info < (3, 6):
-    raise ImportError("Python 3.6 or greater required.")
-else:
-    del sys
-
-from .base import *  # noqa
-from .events import *  # noqa
-from .models import *  # noqa
