--- tests/test_mvc/__init__.py
@@ -1,3 +1,0 @@
-from pytest import importorskip
-
-importorskip("spectate.mvc")
